renames connection to store because I'm the worst
