+++ api/store/connection.go
@@ -1,4 +1,4 @@
-package connection
+package store
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/jackc/pgx/v5/pgconn"
 )
 
-type Database interface {
+type Connection interface {
 	Close()
 	Begin(ctx context.Context) (pgx.Tx, error)
 	Exec(ctx context.Context, sql string, arguments ...interface{}) (pgconn.CommandTag, error)
